backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
celery
redis
pymupdf
minio
qdrant-client
anthropic
openai
python-multipart
pydantic-settings

